Split readv/writev related bindings in different C files
