+++ src/unix/lwt_unix_windows.h
@@ -432,10 +432,5 @@ LWT_NOT_AVAILABLE4(unix_wait4)
 LWT_NOT_AVAILABLE5(unix_bytes_recvfrom)
 LWT_NOT_AVAILABLE1(unix_valid_dir)
 LWT_NOT_AVAILABLE1(unix_invalidate_dir)
-LWT_NOT_AVAILABLE3(unix_writev)
-LWT_NOT_AVAILABLE3(unix_writev_job)
-LWT_NOT_AVAILABLE1(unix_iov_max)
 LWT_NOT_AVAILABLE2(unix_bind_job)
 LWT_NOT_AVAILABLE1(unix_getcwd_job)
-LWT_NOT_AVAILABLE3(unix_readv)
-LWT_NOT_AVAILABLE3(unix_readv_job)
