+++ src/unix/unix_c/unix_writev.c
@@ -1,6 +1,8 @@
 /* OCaml promise library
  * http://www.ocsigen.org/lwt
- * Copyright (C) 2010 Jérémie Dimino
+ * Copyright (C) 2009-2010 Jérémie Dimino
+ *               2009 Mauricio Fernandez
+ *               2010 Pierre Chambart
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -21,17 +23,31 @@
 
 #include "lwt_config.h"
 
-#if defined(LWT_ON_WINDOWS)
+#if !defined(LWT_ON_WINDOWS)
+
+#include <caml/memory.h>
+#include <caml/mlvalues.h>
+#include <caml/unixsupport.h>
+#include <errno.h>
 
 #include "lwt_unix.h"
+#include "unix_readv_writev_utils.h"
 
-LWT_NOT_AVAILABLE1(unix_readable)
-LWT_NOT_AVAILABLE1(unix_writable)
-LWT_NOT_AVAILABLE6(unix_madvise)
-LWT_NOT_AVAILABLE4(unix_mincore)
-LWT_NOT_AVAILABLE1(unix_iov_max)
-LWT_NOT_AVAILABLE3(unix_writev)
-LWT_NOT_AVAILABLE3(unix_writev_job)
-LWT_NOT_AVAILABLE3(unix_readv)
-LWT_NOT_AVAILABLE3(unix_readv_job)
+/* writev primitive for non-blocking file descriptors. */
+CAMLprim value lwt_unix_writev(value fd, value io_vectors, value val_count)
+{
+    CAMLparam3(fd, io_vectors, val_count);
+
+    size_t count = Long_val(val_count);
+
+    /* Assemble iovec structures on the stack. No data is copied. */
+    struct iovec iovecs[count];
+    flatten_io_vectors(iovecs, io_vectors, count, NULL, NULL);
+
+    ssize_t result = writev(Int_val(fd), iovecs, count);
+
+    if (result == -1) uerror("writev", Nothing);
+
+    CAMLreturn(Val_long(result));
+}
 #endif
